refactor: Update CodeEditor component to include value prop
This commit updates the CodeEditor component to include a new prop called 'value'. This prop allows the parent component to control the value of the code editor. The 'value' prop is now passed to the Monaco Editor component and is used to set the initial value of the editor. Additionally, the 'value' prop is also passed to the onChange event handler, ensuring that any changes made in the editor are reflected in the parent component.